Add Dynamic text support to Text and LongText tricks
Text and LongText tricks take only static text param and shows it for the whole life time of the debug drawer. Providing an option to add dynamic text to these tricks will cover more usecases like app status updates, device locale, device timezone etc. New dynamicText param is provided with a default value and so it does not introduce any breaking changes.

For this change to work, debug drawer items need to be updated whenever the drawer is open. It should not cause any performance issue as the debug drawer is already smart enough to calculate the diff to update only the necessary items whenever there is an update.