Fix possibility of ConcurrentModification Exception in NetworkLogManager
Since the log might be called from outside at the same time from different threads the sorting could cause ConcurrentModificationException since it was traversing the list while items might be also added to it.
To ensure the issue is fixed we use synchronization block wherever the entries are modified. That's because we must ensure all the items are added to the list and none is lost. Also it is safe since there cannot be in live/dead lock or other issues since we only have one resource we need to take care of and access.
One Concern could be now that ClearLogs might lock the main thread, however since this tool is for debug purposes and clearLogs will be pressed only by the dev / qa team a small performance hiccup shouldn't be an issue and the impact should be negligible in most cases anyway.
The get entries also had to be synchronized since otherwise the outside could still traverse the list via iterator while the items are being added to the list resulting in the same kind of exception

#32 issue